test: cursor left word delete
